refactor: uniformise type answer (doublons, fix paths)
